requirements.txt
aiohttp==3.6.2
APScheduler==3.6.1
asn1crypto==1.0.1
async-timeout==3.0.1
attrs==19.3.0
baidu-aip==2.2.17.0
bcrypt==3.1.7
bs4==0.0.1
cached-property==1.5.1
cffi==1.12.3
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
colorama==0.4.3
crypto==1.4.1
cryptography==2.7
decorator==4.4.0
deprecation==2.0.6
Django==2.1.8
django-debug-toolbar==2.2
django-shortuuidfield==0.1.3
djangorestframework==3.11.1
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
facebook-wda==0.2.2
fire==0.1.3
future==0.17.1
graphviz==0.13.2
humanize==0.5.1
idna==2.7
jsonpath==0.81
jsonschema==2.6.0
kiwisolver==1.1.0
logzero==1.5.0
lxml==4.3.2
matplotlib==3.0.3
mistune==0.8.4
multi-key-dict==2.0.3
multidict==4.7.4
mysqlclient==1.3.12
Naked==0.1.31
nos-python3-sdk==1.0.2
numpy==1.16.4
objgraph==3.4.1
packaging==19.0
paramiko==2.6.0
pbr==5.4.0
Pillow==5.4.1
progress==1.5
psutil==5.6.7
py==1.8.0
pyasn1==0.4.5
pycparser==2.19
pycrypto==2.6.1
pycryptodomex==3.8.1
pycurl==7.43.0.5
pyee==6.0.0
pyjwkest==1.4.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==3.0.0a2
pyppeteer==0.0.25
python-dateutil==2.8.0
python-jenkins==1.5.0
python-memcached==1.59
pytz==2019.1
PyYAML==3.13
qiniu==7.2.8
redis==3.3.8
requests==2.20.1
retry==0.9.2
ronglian-sms-sdk==1.0
rsa==4.0
shellescape==3.4.1
shodan==1.23.0
shortuuid==1.0.1
six==1.12.0
soupsieve==2.0
South==1.0.2
sqlparse==0.3.0
stevedore==1.30.1
texttable==0.9.1
tornado==6.0.1
tqdm==4.36.1
tzlocal==2.0.0
uiautomator==0.3.6
uiautomator2==0.3.3
urllib3==1.24.3
virtualenv==16.6.2
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
websocket-client==0.56.0
websockets==8.0.2
weditor==0.1.2.dev1
whichcraft==0.5.2
xdot==1.1
XlsxWriter==1.2.9
yarl==1.4.2

django-haystack~=2.8.1
Whoosh~=2.7.4
jieba~=0.42.1